Refactor RNG ownership semantics - Change `stim::TableauSimulator`'s constructor to require a move-reference for its rng - Change `stim::FrameSimulator`'s constructor to require a move-reference for its rng - Change `stim::FrameSimulator`'s RNG from a reference to a normal object owned by the simulator - Replace SHARED_TEST_RNG method with INDEPENDENT_TEST_RNG method - Rewrite all tests to use INDEPENDENT_TEST_RNG, fixing potential overlaps in randomness - Also fix some `pybind11::args args` -> `const pybind11::args &args` warnings - Also fix some int vs size_t comparison warnings
Fixes https://github.com/quantumlib/Stim/issues/353

+++ src/stim/mem/simd_bits_range_ref.inl
@@ -103,7 +103,7 @@ simd_bits_range_ref<W> simd_bits_range_ref<W>::operator<<=(int offset) {
     }
     while (offset >= 64) {
         incoming_word = 0ULL;
-        for (int w = 0; w < num_u64_padded(); w++) {
+        for (size_t w = 0; w < num_u64_padded(); w++) {
             cur_word = u64[w];
             u64[w] = incoming_word;
             incoming_word = cur_word;
@@ -114,7 +114,7 @@ simd_bits_range_ref<W> simd_bits_range_ref<W>::operator<<=(int offset) {
         return *this;
     }
     incoming_word = 0ULL;
-    for (int w = 0; w < num_u64_padded(); w++) {
+    for (size_t w = 0; w < num_u64_padded(); w++) {
         cur_word = u64[w];
         u64[w] <<= offset;
         u64[w] |= incoming_word;
